% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordcloud.R
\name{create_wordcloud}
\alias{create_wordcloud}
\title{Create a wordcloud PNG from word frequencies}
\usage{
create_wordcloud(
  frequencies,
  file_name,
  colors = c("#292f69", "#1396d3", "#7ac143"),
  shape = "circle",
  width = 1992,
  height = 1744
)
}
\arguments{
\item{frequencies}{Word frequencies from get_word_counts()}

\item{file_name}{Output file name}

\item{colors}{Defaults to DP colors}

\item{shape}{Default to circle}

\item{width}{Width of PNG in pixels}

\item{height}{Height of PNG in pixels}
}
\description{
Create a wordcloud PNG from word frequencies
}
