% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TFlist.R
\name{get_TFlist}
\alias{get_TFlist}
\title{Get Transcription Factor List by Species}
\usage{
get_TFlist(species = c("Homo", "Mm", "Rn"))
}
\arguments{
\item{species}{A \code{character} string specifying the species.
Must be one of \code{"Homo"} (human), \code{"Mm"} (mouse), or \code{"Rn"} (rat).}
}
\value{
A \code{data.frame} containing the transcription factor list for the chosen species.
}
\description{
This function retrieves a transcription factor (TF) list for a given species
from the \code{TFLink} dataset included in the MRTools package.
The \code{TFLink.rda} file must be stored in the \code{data/} folder of the package
and contain three data frames: \code{TF_Homo}, \code{TF_Mm}, and \code{TF_Rn}.
}
\examples{
\dontrun{
# Load human TF list
tf_human <- get_TFlist("Homo")
head(tf_human)

# Load mouse TF list
tf_mouse <- get_TFlist("Mm")

# Load rat TF list
tf_rat <- get_TFlist("Rn")
}

}
