% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_counts.R
\name{get_word_counts}
\alias{get_word_counts}
\title{Generate word counts}
\usage{
get_word_counts(
  df,
  column,
  sentiments = NULL,
  exclude_words = "",
  min_count = 1
)
}
\arguments{
\item{df}{A data.frame}

\item{column}{A column name as a string}

\item{sentiments}{An optional data.frame of sentiments from tidytext::get_sentiments("bing")}

\item{exclude_words}{An optional vector of words to exclude from the counts. Stop words (e.g. and, but, or) are handled automatically.}

\item{min_count}{The minimum word frequency to include in the final counts. Useful for pruning wordclouds.}
}
\value{
A data.frame of word counts, optionally by sentiment
}
\description{
Generate word counts
}
