% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_workbook.R
\name{write_workbook}
\alias{write_workbook}
\title{Create an excel workbook from a list of data frames}
\usage{
write_workbook(data_input_list, sheet_names, workbook_name)
}
\arguments{
\item{data_input_list}{A list of data frames}

\item{sheet_names}{A vector of sheet names}

\item{workbook_name}{The desired name of the workbook}
}
\description{
Create an excel workbook from a list of data frames
}
